+++ spec/controllers/forced_authorization_spec.rb
@@ -22,8 +22,16 @@ module Decidim::DecidimAwesome
 
     let(:organization) { create(:organization, available_authorizations: ["dummy_authorization_handler"]) }
     let(:user) { create(:user, :confirmed, organization:) }
-    let(:force_authorization_after_login) { %w(dummy_authorization_handler) }
-    let!(:awesome_config) { create(:awesome_config, organization:, var: :force_authorization_after_login, value: force_authorization_after_login) }
+    let(:authorization_groups) do
+      {
+        "some-group" => {
+          "authorization_handlers" => {
+            "dummy_authorization_handler" => {}
+          }
+        }
+      }
+    end
+    let!(:awesome_config) { create(:awesome_config, organization:, var: :authorization_groups, value: authorization_groups) }
 
     shared_examples "forbids access" do
       it "redirects to the required authorizations page" do