+++ spec/forms/admin/config_form_spec.rb
@@ -38,7 +38,6 @@ module Decidim::DecidimAwesome
         }
       end
       let(:user_timezone) { true }
-      let(:force_authorization_after_login) { ["", "dummy_authorization_handler", "another_dummy_authorization_handler"] }
       let(:force_authorization_with_any_method) { true }
       let(:force_authorization_help_text) do
         { en: "Help text" }
@@ -66,13 +65,12 @@ module Decidim::DecidimAwesome
       describe "valid_keys" do
         let(:attributes) do
           {
-            force_authorization_after_login:,
             force_authorization_help_text_en: "Help text"
           }
         end
 
         it "extracts valid keys from params" do
-          expect(subject.valid_keys).to eq([:force_authorization_after_login, :force_authorization_help_text])
+          expect(subject.valid_keys).to eq([:force_authorization_help_text])
         end
       end
 
@@ -159,30 +157,6 @@ module Decidim::DecidimAwesome
           let(:user_timezone) { false }
 
           it { is_expected.to be_valid }
-        end
-      end
-
-      describe "force authorization after login" do
-        let(:attributes) do
-          {
-            force_authorization_after_login:,
-            force_authorization_with_any_method:,
-            force_authorization_help_text:
-          }
-        end
-
-        it { is_expected.to be_valid }
-
-        context "and force authorization after login is empty" do
-          let(:force_authorization_after_login) { [] }
-
-          it { is_expected.to be_valid }
-        end
-
-        context "and force authorization after login is not a valid handler" do
-          let(:force_authorization_after_login) { %w(invalid_handler another_dummy_authorization_handler) }
-
-          it { is_expected.not_to be_valid }
         end
       end